Update to latest Electron@22.0.0/TypeScript@4.9.4
