+++ .nvmrc
@@ -1,1 +1,1 @@
-16.14.2
+16.17.1
