+++ package.json
@@ -69,7 +69,7 @@
     "@binance-chain/javascript-sdk": "^4.2.0",
     "@cosmjs/stargate": "^0.28.10",
     "@cosmos-client/core": "https://github.com/veado/cosmos-client-ts.git#patch/v0.45.13",
-    "@craco/craco": "^7.0.0-alpha.7",
+    "@craco/craco": "^7.0.0",
     "@devexperts/remote-data-ts": "^2.1.1",
     "@devexperts/rx-utils": "^1.0.0-alpha.14",
     "@devexperts/utils": "^1.0.0-alpha.14",
@@ -111,8 +111,8 @@
     "dayjs": "^1.11.3",
     "dotenv": "^16.0.1",
     "electron-debug": "^3.2.0",
-    "electron-log": "^4.4.7",
-    "electron-updater": "^5.0.1",
+    "electron-log": "^4.4.8",
+    "electron-updater": "^5.3.0",
     "electron-window-state": "^5.0.3",
     "eslint": "^8.23.0",
     "ethers": "^5.6.6",
@@ -171,8 +171,8 @@
     "concurrently": "^7.2.1",
     "cross-env": "^7.0.3",
     "crypto-browserify": "^3.12.0",
-    "electron": "^19.0.15",
-    "electron-builder": "^23.0.3",
+    "electron": "^22.0.0",
+    "electron-builder": "^23.6.0",
     "electron-devtools-installer": "^3.2.0",
     "electron-notarize": "^1.2.1",
     "eslint-config-prettier": "^8.5.0",
@@ -199,7 +199,7 @@
     "ts-jest": "^28.0.4",
     "ts-loader": "^9.3.1",
     "ts-node": "^10.9.1",
-    "typescript": "^4.8.4",
+    "typescript": "^4.9.4",
     "wait-on": "^6.0.1",
     "webpack-cli": "^4.10.0",
     "webpack-node-externals": "^3.0.0"
